+++ kpca_scripts/gen_test_dataset.py
@@ -51,8 +51,8 @@ if __name__ == '__main__':
     
     npoints = 4096*5
     nfiles = 12
-    datset_matrix = np.memmap(f'{dspath}/dataset.dat', mode='r', shape=(3000*nfiles, npoints))
-    datset_theta = np.memmap(f'{dspath}/dataset_theta.dat', mode='r', shape=(3000*nfiles,1))
+    datset_matrix = np.memmap(f'{dspath}/dataset.dat',dtype='float32', mode='r', shape=(3000*nfiles, npoints))
+    datset_theta = np.memmap(f'{dspath}/dataset_theta.dat',dtype='float32', mode='r', shape=(3000*nfiles,1))
 
     k_args = [
         OLD_KWD,
